more noops and better info

+++ src/utils/rename.rs
@@ -37,7 +37,7 @@ pub fn safe_filename(num: u32, artist: &str, title: &str, filetype: &str) -> Str
     )
 }
 
-pub fn rename((src, dest): RenameAction) -> anyhow::Result<bool> {
+pub fn rename((src, dest): RenameAction, noop: bool) -> anyhow::Result<bool> {
     if src == dest {
         Ok(false)
     } else if dest.exists() {
@@ -46,14 +46,16 @@ pub fn rename((src, dest): RenameAction) -> anyhow::Result<bool> {
         if let Some(parent_dir) = dest.parent() {
             if !parent_dir.exists() {
                 println!("Creating {}", parent_dir.display());
-                std::fs::create_dir_all(parent_dir)?;
+                if !noop {
+                    std::fs::create_dir_all(parent_dir)?;
+                }
             }
         }
 
         let src_dir = src.parent().expect("Cannot find parent of src_dir");
         let dest_dir = dest.parent().expect("Cannot find parent of dest_dir");
 
-        let target_to_print = if dest_dir == src_dir {
+        let target_to_print = if dest_dir == src_dir || src_dir.to_string_lossy() == "" {
             dest.file_name().unwrap()
         } else {
             match dest_dir.strip_prefix(src_dir) {
@@ -68,7 +70,9 @@ pub fn rename((src, dest): RenameAction) -> anyhow::Result<bool> {
             target_to_print.to_string_lossy(),
         );
 
-        std::fs::rename(src, dest).map_err(|e| anyhow::anyhow!(e))?;
+        if !noop {
+            std::fs::rename(src, dest).map_err(|e| anyhow::anyhow!(e))?;
+        }
         Ok(true)
     }
 }
